set build and host based on building environment
Use DEB_BUILD_GNU_TYPE DEB_HOST_GNU_TYPE from dpkg-architecture to
set the build and host architectures. This allows building an i386
package set in an i386 chroot running an amd64 kernel. Without this,
the running kernel version is used for detection and tries to build
the x86_64 assembly in spl, which fails with unknown registers since
the actual compiler is i386.